C++ Project for memory allocation from an arena
Hello :=),

Info: Visual Studio 2015 project using C++ unmanaged.

This is a project of mine to create custom allocation from an arena. My
plan is to use this in a game engine once
all features are code complete and tested. I would like to add the
ability to allocate concurrently by creating a
Resource Lifecycle Manager to handle this.

------- Taruchu :=)